requirements.txt
streamlit
faiss-cpu
sentence-transformers
pandas
numpy
google-generativeai
